+++ lib/pagy/calendar/quarter.rb
@@ -11,7 +11,7 @@ class Pagy # :nodoc:
       # The label for any page, with the substitution of the '%q' token
       def label_for(page, opts = {})
         starting_time = starting_time_for(page.to_i)  # page could be a string
-        opts[:format] = (opts[:format] || @vars[:format]).gsub('%q') { (starting_time.month / 4) + 1 }
+        opts[:format] = (opts[:format] || @vars[:format]).gsub('%q') { (starting_time.month / 3.0).ceil }
         localize(starting_time, opts)
       end