+++ lib/pagy/calendar/week.rb
@@ -6,28 +6,23 @@ class Pagy # :nodoc:
     # Calendar week subclass
     class Week < Calendar
       DEFAULT  = { order:  :asc,      # rubocop:disable Style/MutableConstant
-                   format: '%Y-%W',
-                   first_weekday: :sunday }
-      WEEKDAYS = %i[sunday monday tuesday wednesday thursday friday saturday].freeze
+                   format: '%Y-%W' }
 
       protected
 
       # Setup the calendar variables
       def setup_unit_vars
         super
-        raise VariableError.new(self, :first_weekday, "to be in #{WEEKDAYS.inspect}", @vars[:first_weekday]) \
-              unless WEEKDAYS.include?(@vars[:first_weekday])
-
         if @vars[:offset]  # remove in pagy 6
           Warning.warn '[PAGY WARNING] The week :offset variable has been deprecated and will be ignored from pagy 6. ' \
-                       "Set the :first_weekday variable to be one of #{WEEKDAYS.inspect} instead."
-          @vars[:first_weekday] = WEEKDAYS[@vars[:offset]]
+                       "Set the Date.beginning_of_week variable to be one of #{::Date::DAYS_INTO_WEEK.keys.inspect} instead."
+          Date.beginning_of_week = ::Date::DAYS_INTO_WEEK.keys[@vars[:offset]]
         end
-        @initial = @starting.beginning_of_week(@vars[:first_weekday])
-        @final   = @ending.next_week.beginning_of_week(@vars[:first_weekday])
-        @pages   = @last = (@final - @initial).to_i / 1.week
+        @initial = @starting.beginning_of_week
+        @final   = @ending.next_week.beginning_of_week
+        @pages   = @last = (@with_zone ? (@final.time - @initial.time) : (@final - @initial)).to_i / 1.week
         @from    = starting_time_for(@page)
-        @to      = @from + 1.week
+        @to      = @from.next_week
       end
 
       # Starting time for the page